fix: move notificationEvents into events folder
